Remove old parser in favour of new parser (which may be buggy RIP)
